完成了eframe风格的App初始化函数，改善了代码结构
