+++ .gitignore
@@ -2,6 +2,4 @@
 /src/assets
 /.idea
 /baks
-*.png
 .links.json
-storage.yaml
