+++ src/window/glow_app.rs
@@ -1,5 +1,9 @@
 use crate::{event::UserEvent, window::GlutinWindowContext};
 use std::sync::Arc;
+use std::time::Duration;
+
+use crate::window;
+
 
 pub struct GlowApp {
     proxy: winit::event_loop::EventLoopProxy<UserEvent>,
@@ -9,13 +13,17 @@ pub struct GlowApp {
     repaint_delay: std::time::Duration,
     clear_color: [f32; 3],
     window_hidden: bool,
-    update_ui: Box<dyn Fn(&egui::Context) + Send + Sync + 'static>,
+    set_up: Box<dyn Fn(&egui::Context) -> Box<dyn window::App> + Send + Sync + 'static>,
+    update_ui: Option<Box<dyn window::App>>,
+
+    pub winit_window_builder: winit::window::WindowAttributes,
 }
 
 impl GlowApp {
     pub fn new(
+        winit_window_builder: winit::window::WindowAttributes,
         proxy: winit::event_loop::EventLoopProxy<UserEvent>,
-        update_ui: Box<dyn Fn(&egui::Context) + Send + Sync>,
+        set_up: Box<dyn Fn(&egui::Context) -> Box<dyn window::App> + Send + Sync + 'static>,
     ) -> Self {
         Self {
             proxy,
@@ -25,34 +33,49 @@ impl GlowApp {
             repaint_delay: std::time::Duration::MAX,
             clear_color: [0.1, 0.1, 0.1],
             window_hidden: false,
-            update_ui,
+            set_up: set_up,
+            update_ui: None,
+
+            winit_window_builder: winit_window_builder,
         }
     }
-}
 
-fn create_display(
-    event_loop: &winit::event_loop::ActiveEventLoop,
-) -> (GlutinWindowContext, glow::Context) {
-    let glutin_window_context = unsafe { GlutinWindowContext::new(event_loop) };
-    let gl = unsafe {
-        glow::Context::from_loader_function(|s| {
-            let s = std::ffi::CString::new(s)
-                .expect("failed to construct C string from string for gl proc address");
+    fn create_display(
+        &self,
+        event_loop: &winit::event_loop::ActiveEventLoop,
+    ) -> (GlutinWindowContext, glow::Context) {
+        let glutin_window_context = GlutinWindowContext::new(
+            event_loop,
+            self.winit_window_builder.clone(),
+        );
 
-            glutin_window_context.get_proc_address(&s)
-        })
-    };
-
-    (glutin_window_context, gl)
+        let gl = unsafe {
+            glow::Context::from_loader_function(|s| {
+                let s = std::ffi::CString::new(s)
+                    .expect("failed to construct C string from string for gl proc address");
+    
+                glutin_window_context.get_proc_address(&s)
+            })
+        };
+    
+        (glutin_window_context, gl)
+    }
 }
 
 impl winit::application::ApplicationHandler<UserEvent> for GlowApp {
     fn resumed(&mut self, event_loop: &winit::event_loop::ActiveEventLoop) {
-        let (gl_window, gl) = create_display(event_loop);
+        println!("窗体恢复");
+
+        let (gl_window, gl) = self.create_display(event_loop);
         let gl = std::sync::Arc::new(gl);
         gl_window.window().set_visible(true);
 
         let egui_glow = egui_glow::EguiGlow::new(event_loop, gl.clone(), None, None, true);
+
+        // 初始化部分
+
+        self.update_ui = Some(self.set_up.as_mut()(&egui_glow.egui_ctx));
+
 
         let event_loop_proxy = egui::mutex::Mutex::new(self.proxy.clone());
         egui_glow
@@ -78,10 +101,17 @@ impl winit::application::ApplicationHandler<UserEvent> for GlowApp {
             let mut quit = false;
 
             let gl_window = self.gl_window.as_mut().unwrap();
-            self.egui_glow
+
+            if let Some(update_ui) = self.update_ui.as_mut() {
+                self.egui_glow
                 .as_mut()
                 .unwrap()
-                .run(gl_window.window(), self.update_ui.as_mut());
+                .run(gl_window.window(), |ui| {
+                    update_ui.update(ui);
+                });
+            }
+
+            
 
             if quit {
                 event_loop.exit();
@@ -138,7 +168,17 @@ impl winit::application::ApplicationHandler<UserEvent> for GlowApp {
 
         use winit::event::WindowEvent;
         if matches!(event, WindowEvent::CloseRequested | WindowEvent::Destroyed) {
-            event_loop.exit();
+            // if let Some(egui) = self.egui_glow.as_mut() {
+            //     egui.egui_ctx.send_viewport_cmd(egui::ViewportCommand::Visible(false));
+            // }
+            self.proxy
+                .send_event(UserEvent::HideWindow)
+                .unwrap();
+            self.proxy
+                .send_event(UserEvent::Redraw(Duration::ZERO))
+                .unwrap();
+
+            // event_loop.exit();
             return;
         }
